created component for menu bar
