+++ vys-app/src/components/BottomNavBar.tsx
@@ -1,15 +1,18 @@
 'use client';
 
 import { Home, User, Search, Bell, Store } from 'lucide-react';
+import Link from 'next/link';
 
 export default function BottomNavBar() {
   return (
     <nav className="fixed bottom-0 left-0 w-full bg-[#052958] text-[#A1C9FF] border-t border-[#A1C9FF] z-40">
       <ul className="flex justify-around items-center py-3 max-w-screen-md mx-auto">
-        <li className="flex flex-col items-center text-xs sm:text-sm cursor-pointer">
-          <Home className="h-5 w-5 mb-1" />
-          Home
-        </li>
+        <Link href="/">
+          <li className="flex flex-col items-center text-xs sm:text-sm cursor-pointer">
+            <Home className="h-5 w-5 mb-1" />
+            Home
+          </li>
+        </Link>
         <li className="flex flex-col items-center text-xs sm:text-sm cursor-pointer">
           <User className="h-5 w-5 mb-1" />
           Profile
