--- vys-app/src/components/CreateListing.tsx
@@ -1,8 +1,0 @@
-// app/create-listing/page.tsx
-export default function CreateListingPage() {
-  return (
-    <div className="text-white">
-      <h1>Create Listing Page</h1>
-    </div>
-  );
-}
